backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
urllib3==2.2.3
google-generativeai==0.8.3
python-dotenv
tensorflow
opencv-python
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
kaggle
h5py
